Exceptions handling for other models #3005
